requirements.txt
billboard.py==7.1.0
click==8.1.7
Flask==3.1.2
flask_sqlalchemy==3.1.1
lyricsgenius==3.2.0
numpy==2.3.2
PyMySQL==1.0.3
python-dotenv==1.1.1
scikit_learn==1.7.1
